# chronolink-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ChronoLink API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/labels/{timeline}": {
            "get": {
                "tags": [
                    "Label"
                ],
                "summary": "Get all labels",
                "operationId": "labels",
                "parameters": [
                    {
                        "name": "timeline",
                        "in": "path",
                        "description": "Timeline ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of labels",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/create-label": {
            "post": {
                "tags": [
                    "Label"
                ],
                "summary": "Create a label",
                "operationId": "createLabel",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "color",
                                    "timeline_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Label name"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#000000"
                                    },
                                    "timeline_id": {
                                        "type": "string",
                                        "example": "00000000-0000-0000-0000-000000000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Label created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Timeline not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/update-label/{label}": {
            "put": {
                "tags": [
                    "Label"
                ],
                "summary": "Update a label",
                "operationId": "updateLabel",
                "parameters": [
                    {
                        "name": "label",
                        "in": "path",
                        "description": "Label ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "color",
                                    "timeline_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Label name"
                                    },
                                    "color": {
                                        "type": "string",
                                        "example": "#000000"
                                    },
                                    "timeline_id": {
                                        "type": "string",
                                        "example": "00000000-0000-0000-0000-000000000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Label updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Label not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/delete-label/{label}": {
            "delete": {
                "tags": [
                    "Label"
                ],
                "summary": "Delete a label",
                "operationId": "deleteLabel",
                "parameters": [
                    {
                        "name": "label",
                        "in": "path",
                        "description": "Label ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Label deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Label not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get all permissions",
                "operationId": "permissions",
                "responses": {
                    "200": {
                        "description": "List of permissions",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/delete-timeline/{timeline}": {
            "delete": {
                "tags": [
                    "Timeline"
                ],
                "summary": "Delete a timeline",
                "operationId": "destroy",
                "responses": {
                    "200": {
                        "description": "Timeline deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}